backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib
redis
celery[redis]
httpx
python-multipart
psycopg2-binary
phonenumbers
pillow
anthropic
boto3
